+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 + Add a new default embed colour, and add custom colours for each guild as an option
 + Add support for environment variable usage for the `token`, `clientid`, `clientsecret` and `redirecturi` network.yml variables
 + Operators can now include roles as well as users
++ Config files can now be updated programmatically
++ Add the ability to retrieve default values from configs
 
 ## alpha-1.4.1
 ~ Actually made optional plugin.yml parameters optional, in line with the documentation